+++ library/docs/changelog.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 - `LaboratoryActivity.Configuration(storage)` constructor. Use `LaboratoryActivity.Configuration.create(storage)` or `LaboratoryActivity.Configuration.builder()` instead.
 
+### Fixed
+- Warning and error level deprecation on generated feature flags is now correctly suppressed.
+
 ## [0.9.4] - 2020-11-27
 
 ### Added